pyproject.toml
[project]
name = "tech-forum-daily"
version = "0.1.0"
description = "技术论坛日报聚合系统"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "alembic>=1.16.0,<2.0.0",
  "apscheduler>=3.10.4,<4.0.0",
  "beautifulsoup4>=4.13.0,<5.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "feedparser>=6.0.11,<7.0.0",
  "httpx>=0.28.0,<1.0.0",
  "jinja2>=3.1.0,<4.0.0",
  "langchain>=0.3.0,<1.0.0",
  "langchain-openai>=0.2.0,<1.0.0",
  "langdetect>=1.0.9,<2.0.0",
  "pymysql>=1.1.0,<2.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "readability-lxml>=0.8.1,<1.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

